moved local state into Header component
